backend/requirements.txt
flask
sqlalchemy
flask_sqlalchemy
flask_cors
services
datetime
pyjwt
uuid